requirements_cpu.txt
---find-links https://data.pyg.org/whl/torch-2.0.0+cpu.html
awkward==2.2.1
awkward-cpp==15
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
fabio==2023.4.1
filelock==3.12.0
fonttools==4.39.4
h5py==3.8.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lit==16.0.5
llvmlite==0.40.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
pyg-lib==0.2.0+pt20cpu
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
silx==1.1.2
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torch-cluster==1.6.1+pt20cpu
torch-geometric==2.3.1
torch-scatter==2.1.1+pt20cpu
torch-sparse==0.6.17+pt20cpu
torch-spline-conv==1.2.2+pt20cpu
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.2
tzdata==2023.3
uproot==5.0.7
urllib3==2.0.2
zipp==3.15.0
